feat: Add scroll logic
This allows users to know where they are on the page. By listening to scroll events and assigning an active class to the link of the current section in focus, this class is styled accordingly to make it distinct from other links. For smaller screens, a small title changes interactively to the name of the current section in focus